CRM-17464 remove deadlock inducing query
Note this removes an instance where a subquery is used when deleting from the acl_contact_cache - the subquery would lock the entire table
and could be to blame for the failmail we have been seeing

The patch is in 4.7 & involves more limited clearing. The only risk is in deleted contacts
maybe possibly being viewable if the cache is in some way not being managed correctly still
and the calling query does not specify this status adequately.

This code is a bit complex and ... in need of rework but I think scary fallout from this change
is unlikely. (We have deployed this patch to Fuzion customers)

Bug: T119447
Change-Id: I6c430af088cd04e8af71f41da85eeadeccaff185