Partial backport of DB initialisation fixes.
This complements https://gerrit.wikimedia.org/r/256152 in backporting improvements to not calling set utf names too often. The initial backport patch did not include some other changes in core with the result being that the setfactoryclass was being called in the 4.6 code location & the 4.7 one - ie. twice. This fixes that. It is part of the larger patch
: https://github.com/civicrm/civicrm-core/commit/635f0b8655fb0b77ba882cde2081e5055326d175 which is part of a larger refactor. I would not want to backport that patch as a whole as there might be other dependencies in it.

Bug: T116724
Change-Id: Ic7ef3a409081edb865724662b123874879b0920b