CRM-17951 (merged into 4.7) modify transaction view to show the financial account of the transaction
Previously it was showing the financial type of the contribution - but the payment could have a different financial account, although that would not be the norm and for most users no difference would be apparent. I added a group_concat in case there are multiple accounts associated with a transaction now, or in the future.

Note we have separately discussed exposing line item detail - this got stalled on how to make it look tidy in the UI

WMF note - this fits into the refund work in terms of us being able to see properly in the UI what things are coded to.

Bug: T124979
Change-Id: I3b9c7ebf87e5d44ac3adec2c30c4f8dda14d8e25