CRM-18366 Fix contact logging detail to show all related changes & revert all related changes
This report switches from a hard-coded list of contact related tables to considering all tables with an FK to civicrm_contact, including entity_tag
group_subscription, batch, financial_item etc. The revert makes the same switch.

Note that with this change a full subset may still not be retrieved - in order to get that you need to enable the changes from CRM-18193 & apply all patches related to that issue (one will build on this)

Change-Id: I587b9140c188b23dfa8df8ebb70a43bff82cf874