CRM-18409 fix inability to view activity if source contact is deleted
This fix adds the ability to access activities through the api with the
same permission checks as in the BAO if 'id' is passed in (and
contact_id is not for 'supporting legacy stuff reasons').

Note that the api functionality is a good thing, but it is actually
being added primarily for the purposes of being able to add a test fix
for the change.

Upstream PR includes the tests:

https://github.com/civicrm/civicrm-core/pull/8170

Bug: T122519

Change-Id: Ie420ba08e506fab6d7647a4f9ae7315b4abc7dc2