CRM-18193 default log_conn values for interacting with mysql outside of CiviCRM.
In this instance @uniqueID will not be set. We had a go at setting it
within the trigger but that was unreliable. Eventual decision was a
string 'c_' to identify a system connection + a timestamp to hourly
granularity + the connection id. Proposed further work is to
specifically block the revert from the UI for these and add a higher
permission or a 'yes I really really mean it' param to the revert api
for these.

Change-Id: I2ce37d945c86227c35e50e7f33c4254f30d3f975