CRM-18842 Dedupe query: remove OR join in favour of more performant UNION
It probably makes more sense to QA this over on github ... https://github.com/civicrm/civicrm-core/pull/8251/commits

Unions are much faster than OR joins. This change took the length of the query to get the dedupes on a large database from
'as long as it took for the server to fall over' to less than one second on a small group of contacts

This query is only affecting one path - ie Individuals - at the moment as I can only extend that as fast as I can write tests.

Bug: T132396
Change-Id: Ib52441fbbe56b4189765db9ff17847b98917221c