Refreshed 4.6.9 codebase
This is a refresh of the 4.6.9 codebase using the script - which regenerates composer, sql & bower_components

Note the vendor update is a previous security fix where we got the composer update but not the actual updates into our repo. The bower_component updates are possibly due to lack of a version lock? Or perhaps the way the bower was committed in the first place. Note that I observed this discrepancy on bower some months ago. I'm going to generate a version based on our patches
against the latest 4.6 code next & add that as a follow up commit.

This is part of grooming our commits to run them onto 4.7 as part of preparing to upgrade for that.

T133806

Change-Id: I74037cfb133950737579f971311f69598b268c37