Remove wmf EntityTag.getdetails api
Explicitly removing this in preparation for 4.7 update, since I'd rather
we did QA on it's removal than see it quietly dropped. I grepped got
'getdetails' and could not find evidence we still use this api. I think
that code got removed in our last upgrade & the comments seem to bear
that out.

Bug: T133806
Change-Id: I4e65b0914f26a5a1bda6a4e204186a5bc0bfa8be