Revert "Revert "CRM-18744 - Remember quicksearch field selection""
This reverts commit 4e6ef24921734829c87f47a0c4ed29eaef12e077.

Core has proposed an alternate patch to my 'revert CRM-18744 hack-patch'.

I guess we should roll out the core one to the users & see how they find it.

The core fix prompts people if the search type has been cached as something other
than sort_name.

It might be more annoying - so I'd rather roll it out as an improvement on what they
have & if it's still annoying we can fall back on hacking, or working to improve - but I guess
our preference is not to have custom hacks.

Bug: T140293
Change-Id: I2ca3d330396cff6056377014ff9df54782510c99