CRM-18744 - Add prompt text when search term not found
Core has proposed this alternative to hacking out the caching of
search variables.

https://github.com/civicrm/civicrm-core/pull/8764

It does work - although whether it will still be annoying is a TBD.

Since this behaviour is LESS annoying that what is currently deployed
(we haven't deployed the revert yet) it's probably the right thing to deploy.

Bug: T140293

Change-Id: I2852b6cf78499a848bcefb277bd1d0a4fbec920b