Hack out the slow part of the upgrade script.
The 4.7.10 upgrade script has a slow line it in. It alters the index length on the image_url field.

The index IS too long but it's slow to run that change and we would probably make it even shorter
ideally - so let's remove the change for now so the point version upgrade does not require an outage.

Aside from this the upgrade script
- adds fields to a table we don't use (price sets)
- adds a couple of reports (relating to the difference between received date & accounting recognition date)
- adds fields to other tables we don't use (contribution_page, pledge_block)
- adds an option value for outputting docx

Adding fields in the upgrade script will cause them to be added to the log
tables but not the triggers during upgrade. This won't cause breakage - just the data
won't get audit-logged until we update the triggers - which can wait until we tidy up
Rosie's new custom fields.

Change-Id: I65951886625f3b8663073f7418e7b017092847bb