CRM-19197 Stop Merge links from dropping the limit parameter
Note that I decided that we did not need to avoid adding parameters to the url if they were empty.

The code is simpler with &reset=1&rgid=&gid=9 and I think that is equally valid. I had to change
one place in the code to make that work - but it was a change that should have been made.

Bug: T139964
Change-Id: If082f670aaa884c5550f06de4b0b896a80b07b32