Do not give validation error on address with no location_type.
To replicate create a contact with 2 * civicrm_address records for the same location type.

On save load one will lose it's type & then it will fail to save. In the case of 'email' an alternative gets selected.

This is the super quick hack version. I want to follow it with a version that will do clean up rather than just ignore. However, the feeling is ignore is better than cause inconvenience or ask Engage staff to fix at this stage.

Bug: T152426
Change-Id: Ibcf1a34abbaad8e986afef82c4fbf41facad9d7c