Let's create some noise!
These calls should be using the civicrm_api3 wrapper, however, the fix I'm making is not quite the
correct one - the watchdog call is for our debugging only & I'm not quite sure the best bail-out to do when there
is an unexpected error - personally I lean to fail fast & fail hard which is the opposite of the code currently.

This change anticipates I will deploy & then try running the job and monitor pro-actively.

I did have a go at IMAPPING to the mailbox from the server but it didn't work. Perhaps I didn't try
hard enough since presumably it does work if the bounce processor does it.

Note the 'it's already broken' argument is a premise of this patch

CRM-20016

Change-Id: I808bd014373196c6a6385979ee80fb04af4d64fd