CRM-20722: 'Lock wait timeout exceeded' error triggered on smart group cache rebuild
This patch is merged upstream & I am submitting it &
https://github.com/civicrm/civicrm-core/pull/10943

to our codebase.

Doing this
a) syncs us with upstream in the area we are working on
b) adds some code consolidation making it easier for us to tinker with the impact of
cache tables.

Note that this change switches from one sql call to many when clearing groups from
the group_contact_cache table. I'm thinking this is probably not a problem for us as we have
not that many groups (& hopefully we can further reduce them).

Also the apparent COMMIT is a bit of a a myth - CRM_Core_Transaction does weird complex
transaction nesting stuff...

Change-Id: I3d6505b04af1134884f660d71316afb2490221d1