Add ability to find duplicates for selected contacts.
Notes:

- This is not in my mind a final implementation. I wish to merge to core as part of
extracting the code to an extension & version iteration & hence some choices may not
be permanent
- In the extension context the addition of the task may be by hook rather than
adding to the array. I could not see any reason for the weird CONST numbering scheme in core
for the task list & ignored it.
- I wanted to confirm whether it is necessary to extend the task class
which seems unwieldly to me & so have done this just extending CRM_Core_Form.
-It occurs to me there are 2 separate actions we are interested in
  1) select a bunch of contacts & find all the matches for them (based on either
default rule or selecting a rule).
  2) dedupe the contacts selected against each other.

The patch here adds an action to contact search results (Find duplicates for these contacts)
that uses the default rule to find all matches for selected contacts. This is the simpler
option of those listed. Implementing 2 will require some more work on the cacheKey mechanism
since we are trying to by-pass the rules.

Normally I would try to focus on what I can get into core first but in this case
I'm looking for a user win & to explore the code more to plan the next step
upstream, which will involve refactoring.

Bug: T151270

Change-Id: I2da53510e879c6260be8cee4a455d4a0a2d54693