Towards CRM-20155 clean up form code in order to consolidate function use.
PR https://github.com/civicrm/civicrm-core/pull/10890

The code in the Find and Merge Duplicate Contacts form has mystified me for along time. As a step towards extracting
dedupe code into separate extension/s I have gone through & tidied up the finding of duplicates to re-use a function
used elsewhere rather than duplicate it on the form. In the process I tried, and failed, to come up with
a rationale for the duplicate form catching represented in setting form properties.

Note that I made the  parameter explicit. I am expecting this to be a URL paramter in the future
(currently deploying that to our site as such in the context of being able to find matches for specfic contacts.
I felt making it explicit now would aid in not missing instances of it when patching later

Change-Id: I694d71570db9f70406c30769b5ee9457305c25c8