CRM-21109 only clear caches once on cli script, consolidate code
This is a combination of the commits in
https://github.com/civicrm/civicrm-core/pull/10943

Bringing this in does not make any actual change for us. However, it makes it easier for us
to play with the following options

1) setting cache mode to FALSE on a per process basis.
Currently there are statics to try to ensure that caches are flushed
only once per process, but without this the add to group & remove from
group functions bypass them. From a script
CRM_Core_Config::setPermitCacheFlushMode(FALSE);
would prevent cache flushing in the session for group_contact_cache & acl caches until the
script ends or it is retoggled. Also acl caches were not previously subject to the statics.
2)We would consider flushing caches by cron rather than during script runtime. There
is a setting for that & we could extend to cover acl cache. Flushing by cron
permits (optionally) the use of TRUNCATE - which appears to be faster for everyone in the world but us :-)

Change-Id: Ib790939421d3b7ce5d2e258154a8015b8652234f