Remove temp tables pro-actively.
These were the only places I could find that I thought we were likely to hit a moderate number
of times where temp tables are created and not deliberately turned up.

The most prevalent of these was the dedupe one, but that is probably only once per run.

We could merge https://gerrit.wikimedia.org/r/#/c/380667/ to reduce group cache table creation & increase the
group cache time out so we rebuild the groups we do use, but those ARE being dropped

Change-Id: I47a66ed27b36e7f1289f603c9c59db7fade1af72