CRM-21753 add pass-through support for criteria in urls on dedupe pages
This permits urls such as
 civicrm/contact/dedupefind?reset=1&rgid=4&gid=&limit=&criteria={"contact":{"contact_id":205}}&action=update

The contents of $criteria['contact'] are passed to the contact api to get a list of contacts to find possible
matches for (acls are applied when generating the match list to both the matched & match list in the next step).

Note that the urls are urlencoded & hence will have %22 rather than a quote so that feels xss safe but I added some
validation before I figured that part out & I think it still makes sense to keep it.

https://github.com/civicrm/civicrm-core/pull/11658

Change-Id: I22129a0c63da4bdcb0151501e10244762651de95