Stop weird cloning to fix export.
In 4.4 someone (ahem seemingly me!) introduced a bit of a hacky patch to clone the DAO as we iterated through
it. The rationale was that some properties were being set on it between iterations & not later being unset for
the next iteration.

In my testing today simply removing this does not cause the original issue to recur. I tested with
gender_id & also with campaign fields per CRM-14398. My suspicion is subsequent fixes have made this
hack-fix obsolete. I will continue to write a unit test on core

Bug: T187903

Change-Id: Id16febc3a068131dbe07f47f39ee3de45e971e96