Update phpWord per merged to master
This update improves handling of br tags & makes the address block
split over several lines as expected (without it the break does not show).

Note that my QA appeared to show a handful of problems but when I tested
with Leanne over screen share the problems were not present when actual 'MS
Word is used to open the docs rather than open office. I can share these screenshots
with a reviewer

https://github.com/civicrm/civicrm-core/pull/11696

Bug: T187251

Change-Id: I0a62ff74b41eedd26daed4eddeddf930ce939abd