Re-apply lost patch : Port activity tab performance fix
This as a patch we had applied over 5.4 but it got missed in the
re-cut of 5.6 - presumably because it was partially merged

To test compare viewing contact 671068 on staging & on live

upstream PR
Per https://github.com/civicrm/civicrm-core/pull/12557

Note, it seems that the test contact has an activity that is very
slow to render but it is not one of the 'first' activities. This patch stops
activities not being displayed from being rendered (only show first 25).

I'm digging into the slow activity still

Bug: T201008
Bug: T204731
Change-Id: Iabb336153dac47d66b8ca7477b6678d0a4c93bb5