Fix path in settings_location.php.
This got altered back in March to be less server specific, but also broken :-(

CkFinder relies on this when uploading images. It's a known pain point & I have
put some time before into making it more robust but not succeeded.

On staging with this patch I get success on upload to server but not on browse server
but I think that is part of the bigger issue & this fixes as far as it goes

Bug: T202330

Change-Id: I76c227e61d2b86037195f5742d6e02bb7fff63f5