Skip html encoding rich-text fields
e.g try editing the prospect field 'Next Step' on a contact locally in a profile created by
the contact summary editor

This ports this PR https://github.com/civicrm/civicrm-core/pull/12841
in order to save html encoded fields like 'Next Steps' on the prospect set
correctly.

I am not sure if this will be the final patch as I have added comments to the
PR to discuss security implications (if any) - however, I'm not concerned about those
in our case as .... firewall.

Bug: T204179
Change-Id: Id0025eb4a3db815731624c3a052dd3069f21c780