Fix misplaced hyphens in regex
Previous regex was failing under PHP 7.3 with
preg_match(): Compilation failed: invalid range in character class at offset 7

Inside a character range surrounded by [], the hyphen is interpreted
as defining a range, unless it's the first character. Move the
hyphens to the first character in the square brackets where it's
supposed to be interpreted as a hyphen.

https://github.com/civicrm/civicrm-core/pull/13064

Change-Id: I1be37c0f2338eb07d988bc34a941ac1e96593ec7