Fix importing to primary phone / email
Leanne observed that duplicate emails are being created when using import to primary.

I did a big dig into this & the upstream PR has cleanup & unit tests. However, once I made sense of it all
the guts of it turned out to be that the location_type_id was being added to the values array not
the params['email'][x] array. Further down the same function the assignment is correct for address
(which has been working.).

https://github.com/civicrm/civicrm-core/pull/14043

Bug: T219239
Change-Id: I835aaccd2f5584661831c41e2be5a397fb210903