Port https://github.com/civicrm/civicrm-core/pull/16078
This supports updates to the deduper form which in turn relies on afform & this

Allow angular modules to provide dynamic html snippets

hook_angularModules - Accept 'partialsCallback' for dynamic partials

This is similar to the antecedent commit which used a field 'snippets' that
contained a list of HTML files and their content. However, this
uses a callback function instead of providing the full HTML content.

This makes it safer to work with the Angular metadata without needing to
load the content of all partials into memory.

Change-Id: Ia70261b6d84d10302080f363786596af69090b45