Remove unnecessary query when updating org, if no name udpate.
I was seeing recoverable deadlocks on the updateCurrentEmployer when Benevity ran, it doesn't update
organization_name so is fully bypassable

https://github.com/civicrm/civicrm-core/pull/16080

This removes an unnecessary query that can be locking at high volume.

The updateCurrentEmployer function sets the organization_name for any related contacts. We don't need to
do this if we are creating the organization for the first time, or updating the organization but not setting (passing in)
organization_name. We can check & avoid in php.

Also we don't need to check for an existing employer_id unless we are dealing with an Individual record. We can
skip that query for Orgs & Households.

These are cheap queries but at high volume can contribute to locks

Bug: T182148
Change-Id: I0c42fe2da4ca2ff5c05af4d9e0ef57eddace262c