Revert our hack to add an assignee to Contact Deleted by Merge
I think this was in order to allow us to make it easy to navigate from a deleted contact to the kept contact. However,
in upstreaming this wasn't agreed. Instead the Contact.getmergedto api was added which uses the relationship between
the 2 activities. We now have a link that leverages that - so I think we can let  this wmf-patch go

Bug: T244327
Change-Id: I620ba31c473c9f2efa5d0c97b6622812adabe5f5