Fix obscure dedupe scenario where 'bad' location data can overwrite good data
I got this merged to core as https://github.com/civicrm/civicrm-core/pull/17993
in order to simplify it out of our dedupe code (since our address handling is still pretty convoluted & I'm
struggling to get to the nub of our bug.

Our api permits the creation of addresses with no actual address-related data. I'm not sure this is a good
thing but one way or another some addresses got into our database with no useful information and I'm not convinced we
are alone.

When merging it's possible for a null address record to be treated as 'not a conflict' with an address record
with the same location but actual data, and for the null record to be retained. This fixes it such that in that
scenario the retained record is the one with data.

This is pretty obscure and an argument could be made for not fixing it. On the other hand this extends tests on the code
(in addition to the fairly extensive set in testBatchMergesAddresses) and I also did a few tangental cleanups

Bug: T256316
Change-Id: I8b3c2a31ed7a3ecd803dbae73f7e5d184f820f13